Homework6

+++ Homework6/Program.cs
@@ -33,13 +33,13 @@ void CoorIntersectionPoint(double k1, double b1, double k2, double b2){
 }
 
 Console.Write("Input k1: ");
-double k1 = Convert.ToInt32(Console.ReadLine());
+double k1 = Convert.ToDouble(Console.ReadLine());
 Console.Write("Input b1: ");
-double b1 = Convert.ToInt32(Console.ReadLine());
+double b1 = Convert.ToDouble(Console.ReadLine());
 Console.Write("Input k2: ");
-double k2 = Convert.ToInt32(Console.ReadLine());
+double k2 = Convert.ToDouble(Console.ReadLine());
 Console.Write("Input b2: ");
-double b2 = Convert.ToInt32(Console.ReadLine());
+double b2 = Convert.ToDouble(Console.ReadLine());
 
 CoorIntersectionPoint(k1, b1, k2, b2);
